requirements.txt
fastapi
uvicorn
pymongo
motor
pydantic
python-dotenv
setuptools
